Finish the testing for the flappy bird game
