+++ Assets/InitTestScene636574026788811930.unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 358010bad5c364f8b8302a04eb790247
+timeCreated: 1521805879
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
