+++ Assets/InitTestScene636574027522278520.unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 5798f0b4a07e7430d881d3ba0d3c4521
+timeCreated: 1521805952
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
