+++ Assets/InitTestScene636574030414836430.unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 68924b981a5634f1e9230eebef55d699
+timeCreated: 1521806241
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
