+++ Assets/Scenes/WaitingRoom/MiniGames/FlappyBird/Scripts/Bird.cs
@@ -45,7 +45,7 @@ public class Bird : MonoBehaviour
 		// If the bird collides with something set it to dead...
 		isDead = true;
 		//...tell the Animator about it...
-		anim.SetTrigger ("Die");
+		anim.SetTrigger("Die");
 		//...and tell the game control about it.
 		GameControl.instance.BirdDied ();
 	}
