+++ Assets/Scenes/WaitingRoom/MiniGames/FlappyBird/Testing/IntegrationTests.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 3f43030d97ef54427a0e00bf0686a5f5
+folderAsset: yes
+timeCreated: 1521807291
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
