+++ Assets/Scenes/WaitingRoom/MiniGames/FlappyBird/Testing/IntegrationTests/FlappyBirdIntegrationTesting.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 6ae800baeecb64ed295e27b1a7c467ee
+timeCreated: 1521807533
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
