+++ Assets/Scenes/WaitingRoom/MiniGames/FlappyBird/Testing/UnitTests.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: cd502146c7eb34ee19f0eb0a0002f406
+guid: 968e5109651be41f1a95dfb82be76e53
 folderAsset: yes
-timeCreated: 1521475868
+timeCreated: 1521807307
 licenseType: Free
 DefaultImporter:
   externalObjects: {}
