+++ Assets/Scenes/WaitingRoom/MiniGames/FlappyBird/Testing/UnitTests/FlappyBirdUnitTesting.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: df87fbddde2184ead923e3deb28edf38
+timeCreated: 1521807460
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
